pyproject.toml
[project]
name = "audio-transcription-mcp"
version = "0.1.0"
description = "MCP server for audio/video transcription using Groq Whisper and whisperx-mlx"
requires-python = ">=3.11,<3.14"
dependencies = [
    "mcp[cli]>=1.0.0",
    "groq>=0.4.0",
    "youtube-transcript-api>=1.0.0",
    "yt-dlp>=2024.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]
